backend/api/requirements.txt
flask
flask-cors
asgiref
hypercorn
werkzeug
asyncpg
quart
quart-cors
pytest 
pytest-asyncio 
pytest-mock
email-validator
pyjwt
quart-auth
python-dateutil
bcrypt

tensorflow==2.12.0
music21
numpy
pandas
#flask==2.0.1
#flask-cors==3.0.10
#werkzeug==2.0.3
python-dotenv
SQLAlchemy
flask_sqlalchemy